requirements.txt
html2text
googlesearch-python
langchain
beautifulsoup4
google-generativeai
streamlit